BertGNN/requirements.txt
cudatoolkit=10.2.243
dgl-cuda10.1=0.6
ignite=0.4.2
nltk=3.4.5
python=3.6.9
pytorch==1.5.0
scikit-learn=0.22
transformers=4.1.1